Implement map and add locking
The workers can now request task, and complete the Map part of the
MapReduce, writing the result to intermediate files.

I need to think of a better way of looking up the map tasks on the
coordinator, so that the coordinator can mark the right map task as
being done when a worker tells it that it completed a Map.